simple_solutions_stripe/requirements.txt
asgiref==3.7.2
gunicorn==21.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
djangorestframework==3.14.0
idna==3.6
mypy==1.7.1
mypy-extensions==1.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
stripe==7.6.0
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.1.0
